Add CNS for answers
